python/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Google Generative AI
google-generativeai==0.3.2

# Vector Database
pinecone-client==3.0.2

# Observability
langfuse==2.2.5

# Utilities
python-dotenv==1.0.0
